% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redpen.R
\docType{package}
\name{redpen-package}
\alias{redpen}
\alias{redpen-package}
\title{redpen: Check student code in R tutorials}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
